pyproject.toml
[project]
name = "crewai-demo"
version = "0.1.0"
description = "A writing assistant using CrewAI with Google Gemini"
authors = [
    {name = "Your Name",email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai (>=0.11.0)",
    "python-dotenv (>=0.19.0)",
    "openai (>=1.0.0)",
    "google-generativeai (>=0.8.5)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
